+++ connection_properties.go
@@ -305,6 +305,16 @@ var propertyTransactionBatchReadOnly = createConnectionProperty(
 	connectionstate.ContextUser,
 	connectionstate.ConvertBool,
 )
+var propertyTransactionTimeout = createConnectionProperty(
+	"transaction_timeout",
+	"The timeout to apply to all read/write transactions on this connection. "+
+		"Setting the timeout to zero means no timeout.",
+	time.Duration(0),
+	false,
+	nil,
+	connectionstate.ContextUser,
+	connectionstate.ConvertDuration,
+)
 
 // ------------------------------------------------------------------------------------------------
 // Statement connection properties.
@@ -317,6 +327,16 @@ var propertyStatementTag = createConnectionProperty(
 	nil,
 	connectionstate.ContextUser,
 	connectionstate.ConvertString,
+)
+var propertyStatementTimeout = createConnectionProperty(
+	"statement_timeout",
+	"The timeout to apply to all statements on this connection. "+
+		"Setting the timeout to zero means no timeout.",
+	time.Duration(0),
+	false,
+	nil,
+	connectionstate.ContextUser,
+	connectionstate.ConvertDuration,
 )
 
 // ------------------------------------------------------------------------------------------------
@@ -460,6 +480,15 @@ var propertyDisableStatementCache = createConnectionProperty(
 	nil,
 	connectionstate.ContextStartup,
 	connectionstate.ConvertBool,
+)
+var propertyConnectTimeout = createConnectionProperty(
+	"connect_timeout",
+	"The amount of time to wait before timing out when creating a new connection.",
+	0,
+	false,
+	nil,
+	connectionstate.ContextStartup,
+	connectionstate.ConvertDuration,
 )
 
 // Generated read-only properties. These cannot be set by the user anywhere.